Enable eth_wasm_tests * Replace eth::eth_tests::wasm_bindgen_tests into eth::eth_wasm_tests * Fix utxo_tests compilation errors on `cargo test --target wasm32-unknown-unknown --package coins`
