+++ mm2src/coins/eth.rs
@@ -65,6 +65,7 @@ mod web3_transport;
 use self::web3_transport::Web3Transport;
 
 #[cfg(test)] mod eth_tests;
+#[cfg(target_arch = "wasm32")] mod eth_wasm_tests;
 
 /// https://github.com/artemii235/etomic-swap/blob/master/contracts/EtomicSwap.sol
 /// Dev chain (195.201.0.6:8565) contract address: 0xa09ad3cd7e96586ebd05a2607ee56b56fb2db8fd
